auth_server/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
sqlalchemy>=2.0.0
bcrypt>=4.0.0
python-multipart>=0.0.9
PyJWT[crypto]>=2.10.0
cryptography>=42.0.0
httpx>=0.27.0
pytest>=8.0.0
